Move getRecipes function declaration to MealPlans view
