+++ client/components/recipes/RecipesGrid.jsx
@@ -2,46 +2,47 @@ import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import Grid from '@mui/material/Grid2';
 import {
+  Button,
   Card,
   CardMedia,
   CardContent,
   CardActions,
   Typography,
 } from '@mui/material'
+import IngredientList from './IngredientList.jsx';
 
-function RecipesGrid() {
-    // Hold the recipes in state
-    const [recipes, setRecipes] = useState([]);
-    // GET the recipes from the database
-    const getRecipes = () => {
-      console.log('Get recipes invoked');
-      // Make axios GET req to /user/recipes
-      axios.get('/user/recipes')
-        .then(({ data }) => {
-          setRecipes(data);
-      }).catch((err) => {
-          console.error('Error GETting the recipes from the server: ', err);
-      });
-        };
-        // When the component mounts, call getRecipes
-        useEffect(getRecipes, []);
+function RecipesGrid({ recipes }) {
   return (
-    <Grid>
+    <Grid
+      container
+      spacing={2}
+    >
       {recipes.map((recipe) => {
         return (
-          <Card
-            sx={{ maxWidth: 345 }}
+          <Grid
             key={recipe._id}
             >
-          <CardContent>
-            <Typography
-              variant="h6"
-            >
-              {recipe.name}
-            </Typography>
-          </CardContent>
-        </Card>
-        )
+            <Card
+              sx={{ width: 1 }}
+              >
+            <CardContent>
+              <Typography
+                variant="h6"
+              >
+                {recipe.name}
+              </Typography>
+              <Typography
+                variant="subtitle1"
+              >Ingredients
+              </Typography>
+              <IngredientList ingredients={recipe.ingredients}/>
+            </CardContent>
+            <CardActions>
+              <Button>Delete</Button>
+            </CardActions>
+          </Card>
+        </Grid>
+          )
       })}
     </Grid>
   )
