+++ client/views/MealPlans.jsx
@@ -11,10 +11,25 @@ import RecipesGrid from '../components/recipes/RecipesGrid.jsx'
 function MealPlans({ handleThemeChange }) {
   // Set a state value to indicate if a custom recipe/meal is being created
   const [makingRecipe, setMakingRecipe] = useState(false);
+  // Hold the recipes in state
+  const [recipes, setRecipes] = useState([]);
   // Function to toggle makeRecipe to true
   const makeRecipe = () => {
     setMakingRecipe(true);
   }
+  // GET the recipes from the database
+  const getRecipes = () => {
+    console.log('Get recipes invoked');
+    // Make axios GET req to /user/recipes
+    axios.get('/user/recipes')
+      .then(({ data }) => {
+        setRecipes(data);
+      }).catch((err) => {
+          console.error('Error GETting the recipes from the server: ', err);
+      });
+    };
+  // When the component mounts, call getRecipes
+    useEffect(getRecipes, []);
   return (
     <div id="root-app">
       <Navigation handleThemeChange={handleThemeChange}/>
@@ -29,9 +44,12 @@ function MealPlans({ handleThemeChange }) {
       ? (<RecipesForm
           makingRecipe={makingRecipe}
           setMakingRecipe={setMakingRecipe}
+          getRecipes={getRecipes}
           />)
       : <CurrentMealPlans />}
-      <RecipesGrid />
+      <RecipesGrid
+        recipes={recipes}
+      />
     </div>
   );
 }
